backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# 🚀 NEW: Pathway - Live Data Streaming & Memory Layer
# Track-1 "Agentic AI with Live Data" - Core Dependency
pathway>=0.7.0

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# NLP & AI
spacy==3.7.2
transformers==4.35.2
# Use CPU-only PyTorch for compatibility. For GPU: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch==2.1.1
scikit-learn==1.3.2
datasets==4.2.0
huggingface-hub

# LLM Integration
groq>=0.4.0  # Already present, ensuring latest version

# Data Validation & Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
aiofiles==23.2.1
python-magic==0.4.27
httpx==0.25.2

# Rate Limiting
slowapi==0.1.9

# Logging & Monitoring
python-json-logger==2.0.7

# Date & Time
python-dateutil==2.8.2
